pyproject.toml
[project]
name = "profanex"
version = "0.0.1"
description = ""
authors = [{ name = "jyablonski9", email = "jyablonski9@gmail.com" }]
requires-python = ">=3.12, <3.14"
readme = "README.md"
maintainers = [{ name = "jyablonski9", email = "jyablonski9@gmail.com" }]
keywords = [
    "profanity",
    "filtering",
    "text-processing",
    "normalization",
    "fuzzy-matching",
    "censorship",
    "content-moderation",
]
license = "MIT"

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Filters",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
dependencies = ["pyyaml>=6.0.2", "rapidfuzz>=3.13.0"]

[dependency-groups]
dev = ["pytest>=8.4.1", "pytest-cov>=6.2.1"]
local = ["ipykernel>=6.29.5", "mypy>=1.16.1", "ruff>=0.12.3"]

[tool.uv]
default-groups = ["dev"]

[tool.hatch.build]
packages = ["profanex"]

[tool.hatch.build.targets.sdist]
exclude = ["docker/", "tests/", "scripts/", "examples/"]

[tool.hatch.build.targets.wheel]
exclude = ["docker/", "tests/", "scripts/", "examples/"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
]
testpaths = ["tests"]
addopts = "-v"
